Renames and delete redundant folders
