+++ index.html
@@ -1,13 +1,12 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <title>eTools | FAM</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta name="description" content="UNICEF Financial Assurance Module" />
 
-<head>
-  <meta charset="utf-8">
-  <title>eTools | FAM</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <meta name="description" content="UNICEF Financial Assurance Module">
-
-  <!--
+    <!--
       The `<base>` tag below is present to support two advanced deployment options:
       1) Differential serving. 2) Serving from a non-root path.
 
@@ -15,85 +14,84 @@
       a) Add a `basePath` property to the build configuration in your `polymer.json`.
       b) Use the `--base-path` command-line option for `polymer build`.
     -->
-  <base href="/ap/">
+    <base href="/ap/" />
 
-  <link rel="icon" href="assets/images/favicon.ico">
+    <link rel="icon" href="assets/images/favicon.ico" />
 
-  <!-- See https://goo.gl/OOhYW5 -->
-  <link rel="manifest" href="manifest.json">
+    <!-- See https://goo.gl/OOhYW5 -->
+    <link rel="manifest" href="manifest.json" />
 
-  <!-- See https://goo.gl/qRE0vM -->
-  <meta name="theme-color" content="#2E3AA1">
+    <!-- See https://goo.gl/qRE0vM -->
+    <meta name="theme-color" content="#2E3AA1" />
 
-  <!-- Add to homescreen for Chrome on Android. Fallback for manifest.json -->
-  <meta name="mobile-web-app-capable" content="yes">
-  <meta name="application-name" content="eTools - Financial Assurance Module">
+    <!-- Add to homescreen for Chrome on Android. Fallback for manifest.json -->
+    <meta name="mobile-web-app-capable" content="yes" />
+    <meta name="application-name" content="eTools - Financial Assurance Module" />
 
-  <!-- Add to homescreen for Safari on iOS -->
-  <meta name="apple-mobile-web-app-capable" content="yes">
-  <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent">
-  <meta name="apple-mobile-web-app-title" content="eTools - Financial Assurance Module">
+    <!-- Add to homescreen for Safari on iOS -->
+    <meta name="apple-mobile-web-app-capable" content="yes" />
+    <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent" />
+    <meta name="apple-mobile-web-app-title" content="eTools - Financial Assurance Module" />
 
-  <!-- Homescreen icons -->
-  <link rel="apple-touch-icon" href="assets/images/manifest/icon-48x48.png">
-  <link rel="apple-touch-icon" sizes="72x72" href="assets/images/manifest/icon-72x72.png">
-  <link rel="apple-touch-icon" sizes="96x96" href="assets/images/manifest/icon-96x96.png">
-  <link rel="apple-touch-icon" sizes="144x144" href="assets/images/manifest/icon-144x144.png">
-  <link rel="apple-touch-icon" sizes="192x192" href="assets/images/manifest/icon-192x192.png">
+    <!-- Homescreen icons -->
+    <link rel="apple-touch-icon" href="assets/images/manifest/icon-48x48.png" />
+    <link rel="apple-touch-icon" sizes="72x72" href="assets/images/manifest/icon-72x72.png" />
+    <link rel="apple-touch-icon" sizes="96x96" href="assets/images/manifest/icon-96x96.png" />
+    <link rel="apple-touch-icon" sizes="144x144" href="assets/images/manifest/icon-144x144.png" />
+    <link rel="apple-touch-icon" sizes="192x192" href="assets/images/manifest/icon-192x192.png" />
 
-  <!-- Tile icon for Windows 8 (144x144 + tile color) -->
-  <meta name="msapplication-TileImage" content="assets/images/manifest/icon-144x144.png">
-  <meta name="msapplication-TileColor" content="#3f51b5">
-  <meta name="msapplication-tap-highlight" content="no">
+    <!-- Tile icon for Windows 8 (144x144 + tile color) -->
+    <meta name="msapplication-TileImage" content="assets/images/manifest/icon-144x144.png" />
+    <meta name="msapplication-TileColor" content="#3f51b5" />
+    <meta name="msapplication-tap-highlight" content="no" />
 
-  <script>
-    // Load and register pre-caching Service Worker
-    if ('serviceWorker' in navigator) {
-      window.addEventListener('load', function () {
-        navigator.serviceWorker.register('service-worker.js', {
-          scope: '/ap/',
+    <script>
+      // Load and register pre-caching Service Worker
+      if ('serviceWorker' in navigator) {
+        window.addEventListener('load', function () {
+          navigator.serviceWorker.register('service-worker.js', {
+            scope: '/ap/'
+          });
         });
-      });
-    }
+      }
 
-    // Redux assumes `process.env.NODE_ENV` exists in the ES module build.
-    // https://github.com/reactjs/redux/issues/2907
-    window.process = {env: {NODE_ENV: 'production'}};
-  </script>
+      // Redux assumes `process.env.NODE_ENV` exists in the ES module build.
+      // https://github.com/reactjs/redux/issues/2907
+      window.process = {env: {NODE_ENV: 'production'}};
+    </script>
 
-  <!-- Add any global styles for body, document, etc. -->
-  <style>
-    body {
-      margin: 0;
-      font-family: 'Roboto', 'Noto', sans-serif;
-      font-size: 14px;
-      font-weight: 400;
-      line-height: 1.5;
-      min-height: 100vh;
-      background-color: #eeeeee;
-      color: rgba(0, 0, 0, 0.87);
-      -webkit-font-smoothing: antialiased;
-    }
+    <!-- Add any global styles for body, document, etc. -->
+    <style>
+      body {
+        margin: 0;
+        font-family: 'Roboto', 'Noto', sans-serif;
+        font-size: 14px;
+        font-weight: 400;
+        line-height: 1.5;
+        min-height: 100vh;
+        background-color: #eeeeee;
+        color: rgba(0, 0, 0, 0.87);
+        -webkit-font-smoothing: antialiased;
+      }
 
-    * {
-      outline: none;
-    }
-  </style>
-</head>
+      * {
+        outline: none;
+      }
+    </style>
+  </head>
 
-<body>
-  <div id="buildRevNo" style="display: none">revNo</div>
-  <app-shell></app-shell>
-  <noscript>
-    Please enable JavaScript to view this website.
-  </noscript>
-  <!-- Load webcomponents-loader.js to check and load any polyfills your browser needs -->
-  <script src="node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
-  <script src="node_modules/web-animations-js/web-animations-next-lite.min.js"></script>
-  <script src="node_modules/dayjs/dayjs.min.js"></script>
-  <script src="node_modules/dayjs/plugin/utc.js"></script>
+  <body>
+    <div id="buildRevNo" style="display: none;">revNo</div>
+    <app-shell></app-shell>
+    <noscript>
+      Please enable JavaScript to view this website.
+    </noscript>
+    <!-- Load webcomponents-loader.js to check and load any polyfills your browser needs -->
+    <script src="node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+    <script src="node_modules/web-animations-js/web-animations-next-lite.min.js"></script>
+    <script src="node_modules/dayjs/dayjs.min.js"></script>
+    <script src="node_modules/dayjs/plugin/utc.js"></script>
 
-  <script type="module" src="src/elements/core-elements/app-shell-element/app-shell.js"></script>
-</body>
-
+    <script type="module" src="src/elements/app-shell/app-shell.js"></script>
+  </body>
 </html>
