+++ src_ts/elements/app-shell/footer/page-footer.ts
