+++ src_ts/elements/app-shell/sidebar-menu/app-menu.ts
