+++ src_ts/elements/common-elements/follow-up-components/follow-up-main/follow-up-main.ts
@@ -14,8 +14,8 @@ import '../follow-up-actions/follow-up-actions';
 import '../follow-up-financial-findings/follow-up-financial-findings';
 import FollowUpFinancialFindings from '../follow-up-financial-findings/follow-up-financial-findings';
 // eslint-disable-next-line
-import '../../../../elements/pages/spot-checks-page-components/report-page-components/summary-findings-element/summary-findings-element';
-import '../../../../elements/pages/audits-page-components/report-page-components/financial-findings/financial-findings';
+import '../../../pages/spot-checks/report-page-components/summary-findings-element/summary-findings-element';
+import '../../../pages/audits/report-page-components/financial-findings/financial-findings';
 import assign from 'lodash-es/assign';
 import isEmpty from 'lodash-es/isEmpty';
