+++ src_ts/elements/pages/audits/report-page-components/key-internal-controls-weaknesses/key-internal-controls-weaknesses.ts
