+++ src_ts/elements/pages/audits/report-page-components/kicw-risk/kicw-risk.ts
