+++ src_ts/elements/pages/micro-assessments/questionnaire-components/risk-tab/risk-tab-styles.ts
