+++ src_ts/elements/pages/micro-assessments/report-page-components/key-internal-controls-tab-styles.ts
