+++ src_ts/elements/pages/micro-assessments/report-page-components/primary-risk-element.ts
