+++ src_ts/elements/pages/micro-assessments/report-page-components/subject-area-element.ts
