+++ src_ts/elements/pages/spot-checks/report-page-components/internal-controls/internal-controls.ts
