+++ src_ts/elements/pages/spot-checks/report-page-components/summary-findings-element/summary-findings-element.ts
